# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/departments": {
      "post": {
        "operationId": "DepartmentController_create",
        "summary": "",
        "description": "Create new department",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateDepartmentRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DepartmentResponseDto" } } }
          },
          "400": { "description": "Incorrect data." },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Departments"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "DepartmentController_getAll",
        "summary": "",
        "description": "Get all department",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/DepartmentResponseDto" } }
              }
            }
          }
        },
        "tags": ["Departments"],
        "security": [{ "bearer": [] }]
      }
    },
    "/departments/{id}": {
      "get": {
        "operationId": "DepartmentController_getById",
        "summary": "",
        "description": "Find by ID department",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DepartmentResponseDto" } } }
          },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Departments"],
        "security": [{ "bearer": [] }]
      },
      "put": {
        "operationId": "DepartmentController_update",
        "summary": "",
        "description": "Update by ID department",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateDepartmentRequestDto" } } }
        },
        "responses": { "500": { "description": "Internal server error" } },
        "tags": ["Departments"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "DepartmentController_delete",
        "summary": "",
        "description": "Delete department",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "204": { "description": "" } },
        "tags": ["Departments"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/auth/registration": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "",
        "description": "Registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SignUpRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "Successfully created user",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponseDto" } } }
          },
          "400": { "description": "Incorrect registration data." },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Authentication"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "",
        "description": "Login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SignInRequestDto" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Authentication"]
      }
    },
    "/auth/refresh": {
      "get": {
        "operationId": "AuthController_refresh",
        "summary": "",
        "description": "Refresh",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Authentication"],
        "security": [{ "bearer": [] }]
      }
    },
    "/auth/password-change": {
      "post": {
        "operationId": "AuthController_passwordChange",
        "summary": "",
        "description": "Password change",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdatePasswordRequestDto" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Authentication"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/access": {
      "get": {
        "operationId": "AccessController_getAll",
        "summary": "",
        "description": "Get all accesses",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Access"],
        "security": [{ "bearer": [] }]
      }
    },
    "/access/user/{id}": {
      "put": {
        "operationId": "AccessController_updateUser",
        "summary": "",
        "description": "Update user to admin",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Access"],
        "security": [{ "bearer": [] }]
      }
    },
    "/access/admin/{id}": {
      "put": {
        "operationId": "AccessController_downgradeAdmin",
        "summary": "",
        "description": "Downgrade admin to user",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Access"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/jobs": {
      "get": {
        "operationId": "JobController_find",
        "summary": "",
        "description": "Find all jobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/JobResponseDto" } }
              }
            }
          }
        },
        "tags": ["Jobs"],
        "security": [{ "bearer": [] }]
      },
      "post": {
        "operationId": "JobController_create",
        "summary": "",
        "description": "Create new job",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResponseDto" } } }
          }
        },
        "tags": ["Jobs"],
        "security": [{ "bearer": [] }]
      }
    },
    "/jobs/{id}": {
      "put": {
        "operationId": "JobController_update",
        "summary": "",
        "description": "Update job",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobRequestDto" } } }
        },
        "responses": {
          "404": { "description": "There isn't any job with id" },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Jobs"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "JobController_delete",
        "summary": "",
        "description": "Delete logged day",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "" },
          "404": { "description": "There isn't any logged day with this id" }
        },
        "tags": ["Jobs"],
        "security": [{ "bearer": [] }]
      }
    }
  },
  "info": { "title": "BACKEND", "description": "REST API", "version": "1.0.0", "contact": {} },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "DepartmentResponseDto": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
        "required": ["id", "name"]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "type": "string" },
          "skills": { "type": "string" },
          "lastSessionDateTime": { "type": "string" },
          "avatarPath": { "type": "string" },
          "phone": { "type": "string" },
          "department": { "$ref": "#/components/schemas/DepartmentResponseDto" }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt",
          "firstName",
          "lastName",
          "role",
          "skills",
          "lastSessionDateTime",
          "avatarPath",
          "phone",
          "department"
        ]
      },
...
      "CreateDepartmentRequestDto": {
        "type": "object",
        "properties": { "name": { "type": "string", "description": "Department Name", "example": "Design" } },
        "required": ["name"]
      },
...
      "SignUpRequestDto": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "passwordConfirm": { "type": "string" },
          "phone": { "type": "string" }
        },
        "required": ["firstName", "lastName", "email", "password", "passwordConfirm"]
      },
      "SignInRequestDto": {
        "type": "object",
        "properties": { "email": { "type": "string" }, "password": { "type": "string" } },
        "required": ["email", "password"]
      },
      "UpdatePasswordRequestDto": {
        "type": "object",
        "properties": { "oldPassword": { "type": "string" }, "newPassword": { "type": "string" } },
        "required": ["oldPassword", "newPassword"]
      },
...
      "JobResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        },
        "required": ["id", "name", "createdAt", "updatedAt"]
      },
      "JobRequestDto": { "type": "object", "properties": { "name": { "type": "string", "description": "Job name" } } }
    }
  }
}